+++ src/Main/user/DeleteUserWin.java
@@ -89,62 +89,4 @@ public class DeleteUserWin extends JFrame{
         setLocation(900,350);
         pack();
     }
-    /**
-     * @author Foo
-     * @param userName
-     * @exception SQLException , happens if any sql query error happens
-     * This method checks if the user exists in the table for the entered user
-     */
-    private boolean CheckUserSQL(String userName) throws SQLException {
-        User user = new User();
-        boolean existing = false;
-        Statement statement = Server.connection.createStatement();
-        ResultSet resultSet = statement.executeQuery("SELECT userName FROM user");
-        while(resultSet.next()){
-            if(userName.equals(resultSet.getString(1))){
-                existing = true;
-                break;
-            }
-        }
-        statement.close();
-        return existing;
-
-    }
-    /**
-     * @author Foo
-     * @param userName
-     * @exception SQLException , happens if any sql query error happens
-     * This method deletes the user that has been entered into the textfield
-     */
-    private void DeleteUserSQL(String userName) throws SQLException {
-        if(userName != usernamefield.getText()){
-            PreparedStatement deletestatement = Server.connection.prepareStatement("delete from user where userName=?");
-            deletestatement.setString(1,userName);
-            deletestatement.executeQuery();
-            deletestatement.close();
-        }
-        else{
-            JOptionPane.showMessageDialog(null,"why");
-        }
-    }
-    /**
-     * @author Foo
-     * @param userName
-     * @exception SQLException , happens if any sql query error happens
-     * This method deletes every billboard that was created by the deleted user
-     */
-    private void DeleteUserBillboardSQL (String userName) throws SQLException{
-        try{
-            PreparedStatement deletebillboardstatement = Server.connection.prepareStatement(
-                    "delete from billboard where UserName=?"
-            );
-            deletebillboardstatement.setString(1,userName);
-            deletebillboardstatement.executeQuery();
-            deletebillboardstatement.close();
-        }
-        catch(SQLException e){
-            System.out.println();
-        }
-    }
-
 }
